+++ core-ng/src/main/java/core/framework/db/Dialect.java
@@ -4,5 +4,5 @@ package core.framework.db;
  * @author neo
  */
 public enum Dialect {
-    MYSQL, POSTGRESQL
+    MYSQL, POSTGRESQL, HSQL
 }